Fixed #37178 -- Moved MiddlewareMixin out of utils.deprecation.
