+++ django/contrib/sessions/middleware.py
@@ -4,8 +4,8 @@ from importlib import import_module
 from django.conf import settings
 from django.contrib.sessions.backends.base import UpdateError
 from django.contrib.sessions.exceptions import SessionInterrupted
+from django.middleware import MiddlewareMixin
 from django.utils.cache import patch_vary_headers
-from django.utils.deprecation import MiddlewareMixin
 from django.utils.http import http_date
