+++ django/contrib/sites/middleware.py
@@ -1,4 +1,4 @@
-from django.utils.deprecation import MiddlewareMixin
+from django.middleware import MiddlewareMixin
 
 from .shortcuts import get_current_site
