+++ docs/releases/3.1.1.txt
@@ -55,7 +55,7 @@ Bugfixes
   ``TemplateView.get_context_data()`` (:ticket:`31877`).
 
 * Enforced thread sensitivity of the :class:`MiddlewareMixin.process_request()
-  <django.utils.deprecation.MiddlewareMixin>` and ``process_response()`` hooks
+  <django.middleware.MiddlewareMixin>` and ``process_response()`` hooks
   when in an async context (:ticket:`31905`).
 
 * Fixed ``__in`` lookup on key transforms for
