+++ tests/csrf_tests/views.py
@@ -1,9 +1,9 @@
 from django.http import HttpResponse
+from django.middleware import MiddlewareMixin
 from django.middleware.csrf import get_token, rotate_token
 from django.template import Context, RequestContext, Template
 from django.template.context_processors import csrf
 from django.utils.decorators import decorator_from_middleware
-from django.utils.deprecation import MiddlewareMixin
 from django.views.decorators.csrf import csrf_protect, ensure_csrf_cookie
